SA-1619 - Extract AggregatedMetrics component
